requirements.txt
discord.py==2.3.2
discord-py-interactions
 discord-py-slash-command==2.0.0
python-dotenv
aiomysql